+++ package.json
@@ -10,7 +10,7 @@
 		"dev:uws": "cross-env ENV=development npm run start-dev:uws",
 		"build": "npx tsc && npx vite build",
 		"build:esm": "npx tsc && cross-env ESM=true npx vite build",
-		"preview": "cross-env MODE=preview ENV=production npm run start-preview",
+		"preview": "cross-env ENV=production npm run start-preview",
 		"preview:fastify": "cross-env MODE=preview ENV=production npm run start-preview:fastify",
 		"preview:uws": "cross-env MODE=preview ENV=production npm run start-preview:uws",
 		"start-dev": "cross-env MAX_WORKERS=4 node -r sucrase/register server/src/index.ts",
