+++ server/.env
@@ -1,0 +1,1 @@
+DISABLE_SSR_CACHE=true
