+++ server/package.json
@@ -17,6 +17,7 @@
 		"pm2": "^5.3.0",
 		"puppeteer-ssr": "file:./src/puppeteer-ssr",
 		"uWebSockets.js": "uNetworking/uWebSockets.js#v20.39.0",
-		"workerpool": "^8.0.0"
+		"workerpool": "^8.0.0",
+		"dotenv": "^16.3.2"
 	}
 }
