+++ server/src/puppeteer-ssr/index.uws.ts
@@ -1,20 +1,20 @@
 import fs from 'fs'
 import path from 'path'
 import { HttpResponse, TemplatedApp } from 'uWebSockets.js'
-import { ENV, MODE, ENV_MODE, SERVER_LESS } from '../constants'
+import { COOKIE_EXPIRED, ENV, ENV_MODE, MODE, SERVER_LESS } from '../constants'
+import DetectBotMiddle from '../middlewares/uws/DetectBot'
+import DetectDeviceMiddle from '../middlewares/uws/DetectDevice'
+import DetectLocaleMiddle from '../middlewares/uws/DetectLocale'
+import DetectRedirectMiddle from '../middlewares/uws/DetectRedirect'
+import DetectStaticMiddle from '../middlewares/uws/DetectStatic'
 import ServerConfig from '../server.config'
 import { IBotInfo } from '../types'
 import CleanerService from '../utils/CleanerService'
 import Console from '../utils/ConsoleHandler'
-import { CACHEABLE_STATUS_CODE, COOKIE_EXPIRED } from './constants'
+import { CACHEABLE_STATUS_CODE } from './constants'
 import { convertUrlHeaderToQueryString, getUrl } from './utils/ForamatUrl.uws'
 import ISRGenerator from './utils/ISRGenerator.next'
 import SSRHandler from './utils/ISRHandler'
-import DetectLocaleMiddle from '../middlewares/uws/DetectLocale'
-import DetectRedirectMiddle from '../middlewares/uws/DetectRedirect'
-import DetectStaticMiddle from '../middlewares/uws/DetectStatic'
-import DetectBotMiddle from '../middlewares/uws/DetectBot'
-import DetectDeviceMiddle from '../middlewares/uws/DetectDevice'
 
 const COOKIE_EXPIRED_SECOND = COOKIE_EXPIRED / 1000
 const ENVIRONMENT = JSON.stringify({
