+++ server/src/puppeteer-ssr/utils/CacheManager.ts
@@ -10,13 +10,19 @@ import {
 	getFileInfo,
 	setRequestTimeInfo,
 } from './Cache.worker/utils'
+import { DISABLE_SSR_CACHE } from '../constants'
+
+console.log(DISABLE_SSR_CACHE)
 
 const MAX_WORKERS = process.env.MAX_WORKERS
 	? Number(process.env.MAX_WORKERS)
 	: 7
 
+const maintainFile = path.resolve(__dirname, '../../../maintain.html')
+
 const CacheManager = () => {
 	const get = async (url: string) => {
+		if (DISABLE_SSR_CACHE) return
 		const pool = WorkerPool.pool(
 			path.resolve(__dirname, `./Cache.worker/index.${resourceExtension}`),
 			{
@@ -37,6 +43,7 @@ const CacheManager = () => {
 	} // get
 
 	const achieve = async (url: string): Promise<ISSRResult> => {
+		if (DISABLE_SSR_CACHE) return
 		if (!url) {
 			Console.error('Need provide "url" param!')
 			return
@@ -74,6 +81,13 @@ const CacheManager = () => {
 	} // achieve
 
 	const set = async (params: ICacheSetParams) => {
+		if (DISABLE_SSR_CACHE)
+			return {
+				html: params.html,
+				response: maintainFile,
+				status: params.html ? 200 : 503,
+			}
+
 		const pool = WorkerPool.pool(
 			path.resolve(__dirname, `./Cache.worker/index.${resourceExtension}`),
 			{
@@ -94,6 +108,7 @@ const CacheManager = () => {
 	} // set
 
 	const remove = async (url: string) => {
+		if (DISABLE_SSR_CACHE) return
 		const pool = WorkerPool.pool(
 			path.resolve(__dirname, `./Cache.worker/index.${resourceExtension}`),
 			{
